rename neo3d package to blaze3d
