+++ src/client/java/net/neoforged/neoforge/client/blaze3d/GpuDeviceProperties.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.neoforge.client.neo3d;
+package net.neoforged.neoforge.client.blaze3d;
 
 import com.mojang.blaze3d.platform.DepthTestFunction;
 import com.mojang.blaze3d.platform.DestFactor;
