+++ src/client/java/net/neoforged/neoforge/client/blaze3d/validation/ValidationGpuTexture.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.neoforge.client.neo3d.validation;
+package net.neoforged.neoforge.client.blaze3d.validation;
 
 import com.mojang.blaze3d.textures.AddressMode;
 import com.mojang.blaze3d.textures.FilterMode;
